refactor(sell): refactor sell form and input file archtirecture and naming
