+++ src/components/ui/index.tsx
@@ -1,2 +1,2 @@
 export { default as Button } from './Button';
-export { default as FormTextInput } from './FormTextInput';
+export { default as FormTextInput } from './inputs/FormTextInput';
